refactor: bundled unit tests with related files
