+++ lib/decorators/httpCode/index.ts
@@ -1,0 +1,1 @@
+export * from './httpCode.decorator';
