+++ lib/decorators/httpMethod/index.ts
@@ -1,0 +1,1 @@
+export * from './httpMethod.decorators';
