+++ lib/decorators/parameter/index.ts
@@ -1,0 +1,1 @@
+export * from './parameter.decorators';
