+++ lib/decorators/setHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './setHeader.decorator';
