+++ lib/decorators/setHeader/setHeader.decorator.spec.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
 import 'reflect-metadata';
-import { HEADER_TOKEN, SetHeader } from '../../lib/decorators';
+import { HEADER_TOKEN, SetHeader } from './setHeader.decorator';
 
 @SetHeader('X-Api', 'true')
 class Test {
