+++ lib/exceptions/common/BadRequestException.ts
@@ -1,4 +1,4 @@
-import { HttpException } from './HttpException';
+import { HttpException } from '../HttpException';
 
 export class BadRequestException extends HttpException {
   public name = 'BadRequestException';
