+++ lib/exceptions/common/UnauthorizedException.ts
@@ -1,4 +1,4 @@
-import { HttpException } from './HttpException';
+import { HttpException } from '../HttpException';
 
 export class UnauthorizedException extends HttpException {
   public name = 'UnauthorizedException';
