+++ lib/exceptions/index.ts
@@ -1,6 +1,2 @@
 export * from './HttpException';
-export * from './BadRequestException';
-export * from './InternalServerErrorException';
-export * from './NotFoundException';
-export * from './UnauthorizedException';
-export * from './UnprocessableEntityException';
+export * from './common';
