+++ lib/pipes/index.ts
@@ -1,5 +1,5 @@
-export * from './parseBoolean.pipe';
-export * from './parseNumber.pipe';
 export * from './validation.pipe';
-export * from './validateEnum.pipe';
-export * from './parseDate.pipe';
+export * from './parseBoolean';
+export * from './parseNumber';
+export * from './validateEnum';
+export * from './parseDate';
