+++ lib/pipes/parseBoolean/index.ts
@@ -1,0 +1,1 @@
+export * from './parseBoolean.pipe';
