+++ lib/pipes/parseDate/index.ts
@@ -1,0 +1,1 @@
+export * from './parseDate.pipe';
