+++ lib/pipes/parseNumber/index.ts
@@ -1,0 +1,1 @@
+export * from './parseNumber.pipe';
