+++ lib/pipes/validateEnum/index.ts
@@ -1,0 +1,1 @@
+export * from './validateEnum.pipe';
